Rename MarketsPreviewPage into MarketsPage
