+++ src/pages/MarketsPage/components/MarketCard/MarketCard.module.scss
