+++ src/pages/MarketsPage/components/MarketCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MarketCard';
