--- src/pages/MarketsPreviewPage/components/MarketPreviewCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MarketPreviewCard';
