--- src/pages/MarketsPreviewPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MarketsPreviewPage';
