+++ src/pages/index.ts
@@ -4,6 +4,6 @@ export { default as LendPage } from './LendPage';
 export { default as LiquidationsPage } from './LiquidationsPage';
 export { default as LoansPage } from './LoansPage';
 export { default as BorrowPage } from './BorrowPage';
-export { default as MarketsPreviewPage } from './MarketsPreviewPage';
+export { default as MarketsPage } from './MarketsPage';
 export { default as MarketPage } from './MarketPage';
 export { default as PoolsCreationPage } from './PoolsCreationPage';
